feat: Add `safe_mode` to prevent banning active users and notify moderation team (#35)
This patch introduces `safe_mode`, which ensures that active users
matching the ban expressions are not purged immediately. Instead,
a ban request is sent to the moderation team for review. This helps
prevent false positives and ensures that only truly problematic users
are purged.

To enable this feature, set `expressions.safe_mode` to `true` in the
configuration file. Note that the request limit must be at least 4, as
this feature checks if the user is active or not. This adds an extra
layer of safety and oversight to the banning process, making it more
reliable and fair.

Suggested-by: Andreas Shimokawa <shimokawa@fsfe.org>
Signed-off-by: Awiteb <a@4rs.nl>