feat: Don't send alerts of ignored users (#38)
Now we have a global database for the guardian, which was previously
used only for the Matrix bot. It is now in the global scope, with the
default path set to `/app/db.redb`.

This patch ensures that when the moderation team presses the ignore
button or reacts with ignore, the guardian will not resend alerts for
ignored users.

Suggested-by: Andreas Shimokawa <shimokawa@fsfe.org>
Signed-off-by: Awiteb <a@4rs.nl>