change: Move `inactive.check_tokens|check_oauth2` to global scope (#54)
These configurations are used by both the `expressions` and `inactive` sections.
In `expressions`, they are used to check if the user is safe to be purged, and
in `inactive`, they determine if the user is active or not.

Before:
```toml
[inactive]
check_tokens = true
check_oauth2 = true
```

After:
```toml
check_tokens = true
check_oauth2 = true
```

Fixes: #52
BREAKING-CHANGE: Check #54
Signed-off-by: Awiteb <a@4rs.nl>