feat: Lazy purge (#55)
This patch adds the ability for lazy purge, which means the moderation
team doesn't have to worry if they accidentally purge a user. The user
will not be purged immediately but after a certain period, which can be
set in `lazy_purge.purge_after`. This feature can be enabled by setting
`lazy_purge.enabled` to `true`.

After purging a user, the guardian will add an undo button. If one of the
moderators clicks undo, the user will be removed from the purge queue

Suggested-by: otto@codeberg.org
Signed-off-by: Awiteb a@4rs.nl