feat: Check the existing users for the `sus` expressions (#56)
This patch adds `expressions.check_sus_existing_users` to check existing
users against the `sus` expressions.

**Note for big instances:**

It is recommended to run this check once, and then use the new
users fetcher and updated users fetcher. When running this check,
it is advisable to reduce the `req_limit` to around 15 and set the
`req_interval` to 10 seconds. This precaution helps prevent spam in
Matrix and avoids disabling the bot in Telegram due to exceeding the
allowed limit.

Suggested-by: Andreas Shimokawa <shimokawa@fsfe.org>
Signed-off-by: Awiteb <a@4rs.nl>